Migrate codebase to fixed-width stdint.h integer types
Refactor: Complete type system migration for portability and consistency

Changes:
- Replace `int` with `int32_t` throughout codebase
- Replace `long` with `int64_t` for offsets and sizes
- Replace `unsigned char` with `uint8_t` for binary data
- Replace `unsigned int` with `uint32_t` for lengths
- Add `#include <stdint.h>` to all source files
- Add `#include <stddef.h>` for proper type definitions

Files modified:
- include/xxd.h: Added stdint.h header
- include/xxd.in: Converted types, added stdint.h header
- src/xxd.c: Complete type migration, updated function signatures
- src/xxd_embed.c: Added stdint.h header
- example/example.c: Updated main() signature with int32_t

Documentation:
- Moved xxd version history to CHANGELOG.md
- Documented type system changes in changelog
- Removed legacy version comments from source code

Benefits:
- Cross-platform consistency with fixed-width types
- Better portability across 32-bit and 64-bit architectures
- Compliance with C99/C11 standards
- Improved code maintainability and clarity

+++ example/example.c
@@ -1,10 +1,11 @@
 #include "xxd.h"
 
 #include <stdio.h>
+#include <stdint.h>
 
-int main(int argc, char* argv[])
+int32_t main(int32_t argc, char* argv[])
 {
   printf("text.txt\n%s\n", xxd_get("text", NULL, NULL));
   printf("rect.svg\n%s\n", xxd_get("rect", NULL, NULL));
-  return 0;
+  return (int32_t)0;
 }
